requirements.txt
langchain
langchain-community
langchain-groq
openai
faiss-cpu
huggingface-hub
sentence-transformers
python-dotenv
langchain-mistralai